+++ src/main/java/com/example/demo/repository/RoleRepository.java
@@ -6,9 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Role findByRole(String role);
-
-//    @Override
-//    void delete(Role role);
 }
