+++ src/main/java/com/example/demo/repository/UserRepository.java
@@ -6,9 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<Userr, Long> {
+public interface UserRepository extends JpaRepository<Userr, Integer> {
     Userr findByEmail(String email);
-
-//    @Override
-//    void delete(Userr user);
 }
